\name{control}
\alias{control}

\title{Control Variables}
\description{The control variables for \code{ROI.plugin.neos}.}

\arguments{
\item{user}{a character string giving the username.}
\item{email}{a character string giving the email address.}
\item{dry_run}{a logical if \code{TRUE} \pkg{ROI} returns the solver call.}
\item{wait}{a logical indicating whether the R interpreter should wait for the 
            command to finish, or run it asynchronously. If \code{TRUE} 
            \pkg{ROI} returns an object of class \code{"neos_job"}.}
}
